fix: v0.129.0 build infrastructure and file-size lint
- Regenerate SBOM (v0.129.0) to fix SBOM version check failure
- Split json_mapping.rs into main + writeback submodule (542 + 996 lines) to fix 1000-line lint gate
- Moves writeback implementation details to src/json_mapping/writeback.rs while keeping public API in root module
